locate_point_2d: separate fixture, test; + edges
- Migrated fixture to its own header/source files.
- Added new test file to test locate_point_on_edge functionality.
- Fixed nglob assertion in mapping utility to not require inequality for nspec = 1.
